The do_ypcall_tr "resp" parameter is type caddr_t. This necessitates a cast to ypresp_val * to reference resp->status for use as an argument to ypprot_err which provides the return value for do_ypcall_tr.
Since caddr_t and ypresp_val * can have different alignments on some archs such
as arm, GCC -Wcast-align will issue a warning which yp-tools is configured to
treat as an error.

This commit changes the type of the resp parameter of do_ypcall_tr from caddr_t
to ypresp_val * in order to avoid the posibility of an unaligned cast.

Signed-off-by: Jonathan Ben Avraham <yba@tkos.co.il>
Signed-off-by: Thorsten Kukuk <kukuk@thkukuk.de>